Added support for testing cool CPUs, https://github.com/OpenHFT/Chronicle-Core/issues/83

+++ src/main/java/net/openhft/chronicle/bytes/BytesStore.java
@@ -177,13 +177,12 @@ public interface BytesStore<B extends BytesStore<B, Underlying>, Underlying>
         return start() <= offset && offset < safeLimit();
     }
 
-    default boolean inside(long offset, int buffer) {
+    default boolean inside(long offset, long buffer) {
         return start() <= offset && offset + buffer < safeLimit();
     }
 
     /**
-     * @return how many bytes can be safely read, i.e. what is the real capacity of the underlying
-     * data.
+     * @return how many bytes can be safely read, i.e. what is the real capacity of the underlying data.
      */
     default long safeLimit() {
         return capacity();
